fix: refactor export postState file
